bug: docker images don't use init system (#127)
At the moment, the images that are in-use do not actually have
any init system.  The problem with this is that Ansible will keep
spawning defunct processes that won't be getting cleaned up
and over a long Molecule run, it will simply crash / stop responding.

After 20 minutes of a Molecule run, you're up to 72 defunct
processes and it keeps counting.  This is an issue that was
also reported upstream inside Ansible and the resolution is
to use an init system (like tini)[1].

[1]: https://github.com/ansible/ansible/issues/49270#issuecomment-462440983